+++ Simple Game Library/src/at/fhooe/mtd/sgl/graphics/screen/Screen.java
@@ -161,8 +161,10 @@ public class Screen implements Graphics {
         return state.isFullscreen();
     }
 
+    @Override
     public void setTitle(String title) {
-        this.title = title;
+    	this.title = title;
+    	state.setTitle(title);
     }
     
     public String getTitle() {