+++ Simple Game Library/src/at/fhooe/mtd/sgl/graphics/screen/WindowedState.java
@@ -327,4 +327,9 @@ class WindowedState extends ScreenState {
 		return canvas.getLocationOnScreen();
 	}
 
+	@Override
+	public void setTitle(String title) {
+		frame.setTitle(title);
+	}
+
 }